models, serializers and views are split to separate files; drf-yasg is set; service-objects are set; nested serializer for user and gallery_user; update user on serializer level (without updating of 'username'!)
