+++ server/gallery_api/gallery_api/rest_api/models/__init__.py
@@ -1,0 +1,5 @@
+from .comment import Comment
+from .gallery_user import GalleryUser
+from .photo import Photo
+from .trailing_comment import TrailingComment
+from .vote import Vote
