+++ server/gallery_api/gallery_api/rest_api/serializers/login_vk.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+
+class LoginVKSerializer(serializers.Serializer):
+    code = serializers.CharField()
+    network = serializers.CharField()
